# openapi/newton-backend-parity.yaml
openapi: 3.1.0
info:
  title: Newton Backend Parity API
  description: Contract for Rust `newton serve` parity endpoints sourced from Rust DTOs and handlers.
  contact:
    name: Newton Loop Framework
  license:
    name: ''
  version: 0.5.101
servers:
- url: /api/v1
paths:
...
  /workflow-files:
    get:
      tags:
      - workflow-files
      operationId: list_workflow_files
      responses:
        '200':
          description: List of workflow files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowFileSummary'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: File store not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /workflow-files/validate:
    post:
      tags:
      - workflow-files
      operationId: validate_workflow_file
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutWorkflowFileBody'
        required: true
      responses:
        '200':
          description: Diagnostics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowFileDiagnostics'
        '422':
          description: Unparseable YAML
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /workflow-files/{name}:
    get:
      tags:
      - workflow-files
      operationId: get_workflow_file
      parameters:
      - name: name
        in: path
        description: Workflow file name (no .yaml extension)
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Workflow file detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowFileDetail'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: File store not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      tags:
      - workflow-files
      operationId: put_workflow_file
      parameters:
      - name: name
        in: path
        description: Workflow file name (no .yaml extension)
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutWorkflowFileBody'
        required: true
      responses:
        '200':
          description: Updated workflow file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowFileDetail'
        '201':
          description: Created workflow file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowFileDetail'
        '409':
          description: ETag conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Validation error or unparseable YAML
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: File store not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - workflow-files
      operationId: delete_workflow_file
      parameters:
      - name: name
        in: path
        description: Workflow file name (no .yaml extension)
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: File store not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - category
      - message
      properties:
        category:
          type: string
        code:
          type: string
        details: {}
        message:
          type: string
...
    LintFinding:
      type: object
      required:
      - code
      - severity
      - message
      properties:
        code:
          type: string
        location:
          type:
          - string
          - 'null'
        message:
          type: string
        severity:
          type: string
        suggestion:
          type:
          - string
          - 'null'
...
    PutWorkflowFileBody:
      type: object
      required:
      - content
      properties:
        content:
          type: string
        expected_hash:
          type:
          - string
          - 'null'
...
    WorkflowFileDetail:
      type: object
      required:
      - name
      - content
      - content_hash
      - modified_at
      - diagnostics
      properties:
        content:
          type: string
        content_hash:
          type: string
        diagnostics:
          $ref: '#/components/schemas/WorkflowFileDiagnostics'
        modified_at:
          type: string
          format: date-time
        name:
          type: string
    WorkflowFileDiagnostics:
      type: object
      required:
      - parse_ok
      - validate_ok
      - lint
      properties:
        lint:
          type: array
          items:
            $ref: '#/components/schemas/LintFinding'
        parse_error:
          type:
          - string
          - 'null'
        parse_ok:
          type: boolean
        validate_ok:
          type: boolean
    WorkflowFileSummary:
      type: object
      required:
      - name
      - size_bytes
      - modified_at
      - content_hash
      properties:
        content_hash:
          type: string
        description:
          type:
          - string
          - 'null'
        metadata_name:
          type:
          - string
          - 'null'
        modified_at:
          type: string
          format: date-time
        name:
          type: string
        size_bytes:
          type: integer
          format: int64
          minimum: 0
        tags:
          type:
          - array
          - 'null'
          items:
            type: string
...